Revert "Merge branch 'staging' into main"
This reverts commit 28916d277cee700ac204b1b5f96463cfa297b9ef, reversing
changes made to 9acbac4c3d08eb671c3bbd724abd4017c9aa7d53.

+++ temba/classifiers/models.py
@@ -173,7 +173,9 @@ class Classifier(SmartModel):
         on_transaction_commit(lambda: sync_classifier_intents.delay(self.id))
 
     def release(self):
-        assert not self.dependent_flows.exists(), "can't delete classifier currently in use by flows"
+        dependent_flows_count = self.dependent_flows.count()
+        if dependent_flows_count > 0:
+            raise ValueError(f"Cannot delete Classifier: {self.name}, used by {dependent_flows_count} flows")
 
         # delete our intents
         self.intents.all().delete()
